Added README file
Please note the list format of the doc with numbers, letters, and roman numerals. Please follow that format to keep contents organized and easy to read for everyone as we continue to add to it. 

When making changes and committing those changes, please include number list with that line of text in the commit comment. For example: 
"Added number list 2. Interface (UI) and its following sublists."